backend: move clean_game func to helpers module
The main motivation behind this was I wanted to write a test for
the clean_game_entry function, but didn't want to import the populate_db
script, this way is better.
